+++ client/tests/input_selection/alias_foundry_outputs.rs
@@ -20,11 +20,7 @@ const TOKEN_SUPPLY: u64 = 1_813_620_509_061_365;
 
 #[test]
 fn input_selection_alias() -> Result<()> {
-    let rent_structure = RentStructure::build()
-        .byte_cost(500)
-        .key_factor(10)
-        .data_factor(1)
-        .finish();
+    let rent_structure = RentStructure::new(500, 10, 1);
 
     let alias_id_0 = AliasId::from_str("0x0000000000000000000000000000000000000000000000000000000000000000").unwrap();
     let alias_id_1 = AliasId::from_str("0x1111111111111111111111111111111111111111111111111111111111111111").unwrap();