+++ types/tests/rent.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use iota_types::block::{
-    output::{Output, Rent, RentStructure, RentStructureBuilder},
+    output::{Output, Rent, RentStructure},
     protocol::protocol_parameters,
     rand::output::{rand_alias_output, rand_basic_output, rand_foundry_output, rand_nft_output},
 };
@@ -12,11 +12,7 @@ const FACTOR_KEY: u8 = 10;
 const FACTOR_DATA: u8 = 1;
 
 fn config() -> RentStructure {
-    RentStructureBuilder::new()
-        .byte_cost(BYTE_COST)
-        .key_factor(FACTOR_KEY)
-        .data_factor(FACTOR_DATA)
-        .finish()
+    RentStructure::new(BYTE_COST, FACTOR_KEY, FACTOR_DATA)
 }
 
 fn output_in_range(output: Output, range: std::ops::RangeInclusive<u64>) {